glifo

Type Alias NormalizedCoord

pub type NormalizedCoord = i16;

A normalized variation coordinate (for variable fonts) in 2.14 fixed point format.

In most cases, this can be cast from the normalised coords provided by your text layout library.

Equivalent to [skrifa::instance::NormalizedCoord], but defined in Glifo so that Skrifa is not part of Glifo’s public API. This allows Glifo to update its Skrifa in a patch release, and limits the need for updates only to align Skrifa versions.