glifo::util

Trait AffineExt

pub(crate) trait AffineExt {
    // Required methods
    fn has_skew(&self) -> bool;
    fn has_vertical_skew(&self) -> bool;
    fn is_positive_uniform_scale_without_skew(&self) -> bool;
    fn is_positive_uniform_scale_without_vertical_skew(&self) -> bool;
    fn has_non_unit_skew_or_scale(&self) -> bool;
}

Required Methods

fn has_skew(&self) -> bool

Whether the transform has any skewing coefficient.

fn has_vertical_skew(&self) -> bool

Whether the transform has a vertical skew.

fn is_positive_uniform_scale_without_skew(&self) -> bool

Whether the transform has positive, uniform scaling factors and no skew.

fn is_positive_uniform_scale_without_vertical_skew(&self) -> bool

Whether the transform has positive, uniform scaling factors and no vertical skew.

fn has_non_unit_skew_or_scale(&self) -> bool

Whether the transform has non-unit scale or skew.

Note that negative scales (i.e. -1.0) are explicitly allowed.